+++ build.clj
@@ -6,21 +6,53 @@
   For more information, run:
 
   clojure -A:deps -T:build help/doc"
+  (:refer-clojure :exclude [test])
   (:require [clojure.tools.build.api :as b]
-            [org.corfield.build :as bb]))
+            [clojure.tools.deps :as t]
+            [deps-deploy.deps-deploy :as d]))
 
 (def lib 'com.github.seancorfield/clj-new)
 (def version (format "1.2.%s" (b/git-count-revs nil)))
+(def class-dir "target/classes")
+
+(defn test "Run all the tests." [opts]
+  (println "\nRunning tests...")
+  (let [basis    (b/create-basis {:aliases [:test]})
+        combined (t/combine-aliases basis [:test])
+        cmds     (b/java-command
+                  {:basis basis
+                   :java-opts (:jvm-opts combined)
+                   :main      'clojure.main
+                   :main-args ["-m" "cognitect.test-runner"]})
+        {:keys [exit]} (b/process cmds)]
+    (when-not (zero? exit) (throw "Tests failed")))
+  opts)
+
+(defn- jar-opts [opts]
+  (assoc opts
+         :lib lib :version version
+         :jar-file (format "target/%s-%s.jar" lib version)
+         :scm {:tag (str "v" version)}
+         :basis (b/create-basis {})
+         :class-dir class-dir
+         :target "target"
+         :src-dirs ["src"]
+         :src-pom "template/pom.xml"))
 
 (defn ci "Run the CI pipeline of tests (and build the JAR)." [opts]
-  (-> opts
-      (assoc :lib lib :version version)
-      (bb/run-tests) ; there are no tests yet, but that's OK
-      (bb/clean)
-      (assoc :src-pom "template/pom.xml")
-      (bb/jar)))
+  (test opts)
+  (b/delete {:path "target"})
+  (let [opts (jar-opts opts)]
+    (println "\nWriting pom.xml...")
+    (b/write-pom opts)
+    (println "\nCopying source...")
+    (b/copy-dir {:src-dirs ["resources" "src"] :target-dir class-dir})
+    (println "\nBuilding" (:jar-file opts) "...")
+    (b/jar opts))
+  opts)
 
 (defn deploy "Deploy the JAR to Clojars." [opts]
-  (-> opts
-      (assoc :lib lib :version version)
-      (bb/deploy)))
+  (let [{:keys [jar-file] :as opts} (jar-opts opts)]
+    (d/deploy {:installer :remote :artifact (b/resolve-path jar-file)
+               :pom-file (b/pom-path (select-keys opts [:lib :class-dir]))}))
+  opts)