+++ src/clj/new/polylith/project_app_build.clj
@@ -1,13 +1,28 @@
 (ns build
   (:refer-clojure :exclude [test])
-  (:require [org.corfield.build :as bb]))
+  (:require [clojure.tools.build.api :as b]))
 
 (def lib '{{group}}/{{artifact}})
 (def version "{{version}}")
 (def main '{{namespace}}.cli.main)
+(def class-dir "target/classes")
+
+(defn- uber-opts [opts]
+  (assoc opts
+         :lib lib :main main
+         :uber-file (format "target/%s-%s.jar" (name lib) version)
+         :basis (b/create-basis {})
+         :class-dir class-dir
+         :src-dirs ["src"]
+         :ns-compile [main]))
 
 (defn uber "Build the uberjar." [opts]
-  (-> opts
-      (assoc :lib lib :version version :main main)
-      (bb/clean)
-      (bb/uber)))
+  (b/delete {:path "target"})
+  (let [opts (uber-opts opts)]
+    (println "\nCopying source...")
+    (b/copy-dir {:src-dirs ["resources" "src"] :target-dir class-dir})
+    (println (str "\nCompiling " main "..."))
+    (b/compile-clj opts)
+    (println "\nBuilding JAR...")
+    (b/uber opts))
+  opts)